move user register from user controller to auth controller
