+++ service/src/tests/controllers/auth.test.js
@@ -13,6 +13,15 @@ afterAll(async () => {
 });
 
 describe('Auth controller test', () => {
+    it('#Test register method; POST /api/auth/register; ', async () => {
+        const userData = {...devUsers[0], email: 'xhc044x@gmail.com'};
+        const response = await request(app)
+            .post('/api/auth/register')
+            .send(userData);
+
+        expect(response.body.success).toBe(true);
+    });
+    
     it('#Test signin method; POST /api/auth/signin; ', async () => {
         const userData = { email: devUsers[0].email, password: devUsers[0].password };
         const response = await request(app)
